Module vmem::vmem

Vmem Resource Allocator

Vmem is a general-purpose resource allocator which provides constant-time performance with low fragmentation.

For a proper explanation of the allocator, see the paper Magazines and Vmem: Extending the Slab Allocator to Many CPUs and Arbitrary Resources.

Allocation Strategies

Each allocation request can specify an allocation strategy.

See the documentation for AllocStrategy for information on the individual allocation strategies available.

Structs

Error

Layout

Layout of a constrained allocation

Vmem

Vmem Arena

Enums

AllocStrategy

Allocation Strategy

Traits

Source

Types from which a Vmem arena can import spans.

Functions

bootstrap⚠

Bootstrap the vmem system

Type Definitions

Result