Struct vmem::vmem::Vmem

pub struct Vmem<'name, 'src> { /* private fields */ }

Vmem Arena

See the module-level documentation for more information.

Implementations

impl<'name, 'src> Vmem<'name, 'src>

pub const fn new( name: Cow<'name, str>, quantum: usize, source: Option<&'src dyn Source> ) -> Self

Create a new arena

Panics

quantum must be a power-of-two.

pub fn with_span( name: Cow<'name, str>, quantum: usize, source: Option<&'src dyn Source>, base: usize, size: usize ) -> Result<Self>

Create a new arena with an initial span

Errors

This function errors only if resources cannot be allocated to describe the new span.

Panics

quantum must be a power-of-two.

impl Vmem<'_, '_>

pub fn name(&self) -> &str

Returns the name of this arena.

pub fn capacity(&self) -> usize

Returns the total number of bytes managed by the arena

pub fn capacity_used(&self) -> usize

Returns the total number of bytes currently allocated from the arena

pub fn capacity_free(&self) -> usize

Returns the total number of bytes in the arena available for allocation

pub fn add(&self, base: usize, size: usize) -> Result<()>

Add a span to this arena

Errors

This function errors only if resources cannot be allocated to describe the new span.

pub fn alloc(&self, size: usize, strategy: AllocStrategy) -> Result<usize>

Allocate size bytes from this arena.

Errors

This function errors if the requested allocation cannot be satisfied by this arena.

pub unsafe fn free(&self, base: usize, size: usize)

Free a segment allocated by alloc()

Safety

The segment must have previously been allocated by a call to alloc().

Panics

This function panics if the segment cannot be found in the allocation hash table.

pub fn alloc_constrained( &self, layout: Layout, strategy: AllocStrategy ) -> Result<usize>

Allocate a segment which respects the constraints of layout.

Errors

This function errors if the requested allocation cannot be satisfied by this arena.

pub unsafe fn free_constrained(&self, base: usize, size: usize)

Free a segment allocated by alloc_constrained()

Safety

The segment must have previously been allocated by a call to alloc_constrained().

Panics

This function panics if a matching segment cannot be found in the allocation table.