Trait vmem::vmem::Source

pub unsafe trait Source: Send + Sync {
    fn import(&self, size: usize) -> Result<usize>;
    unsafe fn release(&self, base: usize, size: usize);
}

Types from which a Vmem arena can import spans.

Safety

Required Methods

fn import(&self, size: usize) -> Result<usize>

Import a span from this source.

Errors

Returns Err if no resources are currently available.

unsafe fn release(&self, base: usize, size: usize)

Release a segment back to its source

Safety

The segment base .. base + size must have been previously allocated by a call to import() on the same source.